split users to two components
